requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pymongo
groq
python-dotenv
requests